% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{cluster_oslom}
\alias{cluster_oslom}
\title{OSLOM Community Detection}
\usage{
cluster_oslom(edgelist, oslom_bin_undir_path, seed)
}
\arguments{
\item{edgelist}{An igraph edgelist object}

\item{oslom_bin_undir_path}{The path to oslom undirected network binary}

\item{seed}{An integer specifying the seed to replicate the result}
}
\description{
Wrapper for OSLOM Community Detection \url{http://oslom.org/}
}
\references{
Finding statistically significant communities in networks A. Lancichinetti, F. Radicchi, J.J. Ramasco and S. Fortunato PLoS ONE 6, e18961 (2011).
}
