% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.R
\name{create_n_runs_warp_lda}
\alias{create_n_runs_warp_lda}
\title{WarpLDA Topic Modelling}
\usage{
create_n_runs_warp_lda(
  k,
  n_runs,
  alpha,
  beta,
  dtm,
  lda_params,
  store_path = NULL
)
}
\arguments{
\item{k}{An integer specifying the number of topics}

\item{n_runs}{The nuber of runs for LDA}

\item{alpha}{Alpha parameter as specified in WarpLDA}

\item{beta}{Beta parameter as specified in WarpLDA}

\item{dtm}{The document term matrix representation of the corpus}

\item{lda_params}{Additional parameters for WarpLDA}

\item{store_path}{Optional path to serialize topic modelling result}
}
\description{
WarpLDA Topic Modelling
}
\references{
Chen, J., Li, K., Zhu, J., & Chen, W. (2016). WarpLDA: a Cache Efficient O(1) Algorithm for Latent Dirichlet Allocation. In VLDB.
}
