% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic.R
\name{find_optimal_topic_parameters}
\alias{find_optimal_topic_parameters}
\title{Find Optimal Topic Parameters}
\usage{
find_optimal_topic_parameters(corpus, lower, upper, stop_words)
}
\arguments{
\item{corpus}{The corpus used for WarpLDA}

\item{lower}{The lower bound for k, alpha and beta (a 3-element numeric vector)}

\item{upper}{The upper bound for k, alpha and beta (a 3-element numeric vector)}

\item{stop_words}{The list of stop words to be used}
}
\value{
Returns a list with the optimal hyperparameters
}
\description{
Find Optimal Topic Parameters
}
\references{
Mika V. Mantyla, Maelick Claes, and Umar Farooq. 2018. Measuring LDA Topic Stability from Clusters of Replicated Runs. In ACM / IEEE International Symposium on Empirical Software Engineering and Measurement (ESEM) (ESEM ’18), October 11–12, 2018, Oulu, Finland. ACM, New York, NY, USA, Article 4, 4 pages. https://doi.org/10.1145/3239235.3267435
}
