% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{parse_record}
\alias{parse_record}
\title{Parse ASRS records obtained from online query}
\usage{
parse_record(folder_path)
}
\arguments{
\item{folder_path}{The folder path where ASRS records obtained from queries are stored.}
}
\value{
A list of tables, each containing one month of reports.
}
\description{
Adjust record headers into single line headers, and split monthly
records.
}
