% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{write_record}
\alias{write_record}
\title{Save parsed ASRS records obtained from online query}
\usage{
write_record(monthly_reports_list, folder_path)
}
\arguments{
\item{monthly_reports_list}{A R list of monthly reports as obtained by \code{parse_record}.}

\item{folder_path}{The folder path where ASRS records obtained from queries are stored.}
}
\value{
No value is returned, and list of monthly reports is saved to `folder_path`/
}
\description{
Save parsed ASRS records obtained from online query
}
